+++ Sources/code/application/Models/Content/PhonePropertiesContentAppModel.swift
@@ -14,6 +14,8 @@ public extension ApplicationClient.Content {
         
         public var number: String?
         
+        public var phoneType: String?
+        
 
         public enum CodingKeys: String, CodingKey {
             
@@ -23,15 +25,19 @@ public extension ApplicationClient.Content {
             
             case number = "number"
             
+            case phoneType = "phone_type"
+            
         }
 
-        public init(code: String? = nil, key: String? = nil, number: String? = nil) {
+        public init(code: String? = nil, key: String? = nil, number: String? = nil, phoneType: String? = nil) {
             
             self.key = key
             
             self.code = code
             
             self.number = number
+            
+            self.phoneType = phoneType
             
         }
 
@@ -74,6 +80,18 @@ public extension ApplicationClient.Content {
             }
             
             
+            
+            do {
+                phoneType = try container.decode(String.self, forKey: .phoneType)
+            
+            } catch DecodingError.typeMismatch(let type, let context) {
+                print("Type '\(type)' mismatch:", context.debugDescription)
+                print("codingPath:", context.codingPath)
+            } catch {
+                
+            }
+            
+            
         }
         
         public func encode(to encoder: Encoder) throws {
@@ -89,6 +107,10 @@ public extension ApplicationClient.Content {
             
             
             try? container.encodeIfPresent(number, forKey: .number)
+            
+            
+            
+            try? container.encodeIfPresent(phoneType, forKey: .phoneType)
             
             
         }