+++ Sources/code/platform/Models/Webhook/SubscriberConfigPostRequestV2WebhookPlatformModel.swift
@@ -12,7 +12,7 @@ public extension PlatformClient.Webhook {
     class SubscriberConfigPostRequestV2: Codable {
         
         
-        public var name: String?
+        public var name: String
         
         public var webhookUrl: String?
         
@@ -28,7 +28,7 @@ public extension PlatformClient.Webhook {
         
         public var authMeta: AuthMeta?
         
-        public var events: [Events]?
+        public var events: [Events]
         
 
         public enum CodingKeys: String, CodingKey {
@@ -53,7 +53,7 @@ public extension PlatformClient.Webhook {
             
         }
 
-        public init(association: Association, authMeta: AuthMeta? = nil, customHeaders: [String: Any]? = nil, emailId: String, events: [Events]? = nil, name: String? = nil, provider: String, status: SubscriberStatus, webhookUrl: String? = nil) {
+        public init(association: Association, authMeta: AuthMeta? = nil, customHeaders: [String: Any]? = nil, emailId: String, events: [Events], name: String, provider: String, status: SubscriberStatus, webhookUrl: String? = nil) {
             
             self.name = name
             
@@ -79,16 +79,9 @@ public extension PlatformClient.Webhook {
             let container = try decoder.container(keyedBy: CodingKeys.self)
             
             
-                do {
-                    name = try container.decode(String.self, forKey: .name)
+                name = try container.decode(String.self, forKey: .name)
                 
-                } catch DecodingError.typeMismatch(let type, let context) {
-                    print("Type '\(type)' mismatch:", context.debugDescription)
-                    print("codingPath:", context.codingPath)
-                } catch {
-                    
-                }
-                
+            
             
             
                 do {
@@ -147,16 +140,9 @@ public extension PlatformClient.Webhook {
                 
             
             
-                do {
-                    events = try container.decode([Events].self, forKey: .events)
+                events = try container.decode([Events].self, forKey: .events)
                 
-                } catch DecodingError.typeMismatch(let type, let context) {
-                    print("Type '\(type)' mismatch:", context.debugDescription)
-                    print("codingPath:", context.codingPath)
-                } catch {
-                    
-                }
-                
+            
             
         }